Notifications fix
